Feat : " Data 수정 및 추가 "
Data
- st_ItemData
 -  ItemCategory -> en_LargeItemCategory 로 수정
 -  en_MediumItemCategory 추가
 -  ItemType -> en_SmallItemCategory 로 수정
- st_CraftingMaterialItemData, st_CraftingCompleteItemData, st_CraftingItemCategoryData
 - en_ItemType -> en_SmallItemCategory로 수정, en_ItemCategory -> en_LargeItemCategory로 수정

DataManager
- LargeCategory, MediumCategory, SmallCategory 파싱 추가

ItemData.json
- 아이템에 LargeCategory, MediumCategory, SmallCategory 추가

EnvironmentData.json, MonsterData.json
- 오브젝트의 드랍 테이블 id 수정 ItemDataSheetId -> DropItemSmallCategory

+++ GameServer/DataManager.h
@@ -8,12 +8,12 @@
 class CDataManager
 {
 public:
-	map<int32, st_ItemData*> _Items;
+	map<int16, st_ItemData*> _Items;
 	map<int16, st_PlayerStatusData*> _Status;
 	map<int32, st_MonsterData*> _Monsters;
 	map<int32, st_SkillData*> _Skills;
 	map<int32, st_EnvironmentData*> _Environments;
-	map<int32, st_CraftingItemCategoryData*> _CraftingData;
+	map<int8, st_CraftingItemCategoryData*> _CraftingData;
 
 	CDataManager()
 	{
